# PatchDb.Backend/PatchDb.Backend.Service/User/UserController.cs
using Amazon.S3.Model.Internal.MarshallTransformations;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;
using PatchDb.Backend.Core.Authentication;
using PatchDb.Backend.Service.User.Models;
using PatchDb.Backend.Service.User.Models.Dto;
using Platform.Core.Application.Persistence;

namespace PatchDb.Backend.Service.User;

[ApiController]
[Route("api/user")]
public class UserController : ControllerBase
{
    private readonly IUserService _userService;

    public UserController(IUserService userService)
    {
        _userService = userService;
    }

    [HttpGet]
    [Authorize]
    public async Task<ActionResult<UserResponse>> GetUser()
        => Ok(await _userService.GetUserById(User.UserId()));

    [HttpGet("{userId}")]
    [Authorize]
    public async Task<ActionResult<UserResponse>> GetUserById(Guid userId)
        => Ok(await _userService.GetUserById(userId, hidePii: true));

    [HttpPut("profile-picture")]
    [Authorize]
    public async Task<ActionResult<UserResponse>> UpdateProfilePicture([FromBody] UpdateProfilePictureRequest request)
        => Ok(await _userService.UpdateProfilePicture(User.UserId(), request.FileId));

    [HttpDelete("profile-picture")]
    [Authorize]
    public async Task<ActionResult<UserResponse>> RemoveProfilePicture()
        => Ok(await _userService.RemoveProfilePicture(User.UserId()));

    [HttpPut("bio")]
    [Authorize]
    public async Task<ActionResult<UserResponse>> UpdateBio([FromBody] UpdateBioRequest request)
        => Ok(await _userService.UpdateBio(User.UserId(), request.Bio ?? string.Empty));

    [HttpGet("all")]
    [Authorize(Roles = nameof(UserRole.Admin))]
    public async Task<ActionResult> GetAllUsers()
        => Ok(await _userService.GetAllUsers());

    [HttpPut("university-information")]
    [Authorize]
    public async Task<ActionResult<UserResponse>> UpdateUniversityInfo([FromBody] UpdateUserUniversityInfoRequest request)
        => Ok(await _userService.UpdateUniversityInfo(User.UserId(), request));

    [HttpPost("search")]
    [Authorize]
    public async Task<ActionResult<List<UserResponse>>> SearchUser([FromBody] SearchUserRequest request)
        => Ok(await _userService.SearchUser(request));
}