# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "patchdb-backend-service",
    "version": "v1"
  },
  "paths": {
...
    "/api/patches/{patchNumber}": {
      "get": {
        "tags": [
          "Patch"
        ],
        "parameters": [
          {
            "name": "patchNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PatchResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatchResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatchResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/user/{userId}": {
      "get": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/profile-picture": {
      "put": {
        "tags": [
          "User"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfilePictureRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfilePictureRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfilePictureRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "User"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/bio": {
      "put": {
        "tags": [
          "User"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBioRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBioRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBioRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/user/university-information": {
      "put": {
        "tags": [
          "User"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserUniversityInfoRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserUniversityInfoRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserUniversityInfoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/search": {
      "post": {
        "tags": [
          "User"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchUserRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchUserRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SearchUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/user-patches/upload/{fileId}": {
      "post": {
        "tags": [
          "UserPatch"
        ],
        "parameters": [
          {
            "name": "fileId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PatchUploadResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatchUploadResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatchUploadResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/user-patches/{userPatchUploadId}/matching-patch-number/{matchingPatchNumber}": {
      "patch": {
        "tags": [
          "UserPatch"
        ],
        "parameters": [
          {
            "name": "userPatchUploadId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "matchingPatchNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserPatchModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPatchModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPatchModel"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/user-patches/{userId}": {
      "get": {
        "tags": [
          "UserPatch"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetUserPatchesResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUserPatchesResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUserPatchesResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "GetUserPatchesResponse": {
        "type": "object",
        "properties": {
          "patches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserPatchModel"
            },
            "nullable": true
          },
          "unmatchesPatches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserPatchUploadModel"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "NewMatchingPatchesModel": {
        "required": [
          "imageUrl"
        ],
        "type": "object",
        "properties": {
          "patchNumber": {
            "type": "integer",
            "format": "int32"
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "patchMaker": {
            "type": "string",
            "nullable": true
          },
          "university": {
            "$ref": "#/components/schemas/UniversityModel"
          },
          "universitySection": {
            "type": "string",
            "nullable": true
          },
          "releaseDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "patchSubmissionId": {
            "type": "string",
            "format": "uuid"
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "updated": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "similarity": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "OwnedMatchingPatchesModel": {
        "required": [
          "matchingPatch",
          "uploads"
        ],
        "type": "object",
        "properties": {
          "userPatchId": {
            "type": "string",
            "format": "uuid"
          },
          "matchingPatch": {
            "$ref": "#/components/schemas/PatchResponse"
          },
          "uploads": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserPatchUploadModel"
            },
            "nullable": true
          },
          "isFavorite": {
            "type": "boolean"
          },
          "aquiredAt": {
            "type": "string",
            "format": "date-time"
          },
          "similarity": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "PatchResponse": {
        "required": [
          "imageUrl"
        ],
        "type": "object",
        "properties": {
          "patchNumber": {
            "type": "integer",
            "format": "int32"
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "patchMaker": {
            "type": "string",
            "nullable": true
          },
          "university": {
            "$ref": "#/components/schemas/UniversityModel"
          },
          "universitySection": {
            "type": "string",
            "nullable": true
          },
          "releaseDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "patchSubmissionId": {
            "type": "string",
            "format": "uuid"
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "updated": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PatchUploadResponse": {
        "required": [
          "upload"
        ],
        "type": "object",
        "properties": {
          "ownedMatchingPatches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OwnedMatchingPatchesModel"
            },
            "nullable": true
          },
          "newMatchingPatches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NewMatchingPatchesModel"
            },
            "nullable": true
          },
          "upload": {
            "$ref": "#/components/schemas/UserPatchUploadModel"
          }
        },
        "additionalProperties": false
      },
...
      "SearchUserRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "nullable": true
          },
          "universityCode": {
            "type": "string",
            "nullable": true
          },
          "skip": {
            "type": "integer",
            "format": "int32"
          },
          "take": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "UniversityModel": {
        "required": [
          "code",
          "logoUrl",
          "name"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "logoUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdateBioRequest": {
        "type": "object",
        "properties": {
          "bio": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdateProfilePictureRequest": {
        "type": "object",
        "properties": {
          "fileId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "UpdateUserUniversityInfoRequest": {
        "type": "object",
        "properties": {
          "universityCode": {
            "type": "string",
            "nullable": true
          },
          "universityProgram": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UserPatchModel": {
        "required": [
          "matchingPatch",
          "uploads"
        ],
        "type": "object",
        "properties": {
          "userPatchId": {
            "type": "string",
            "format": "uuid"
          },
          "matchingPatch": {
            "$ref": "#/components/schemas/PatchResponse"
          },
          "uploads": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserPatchUploadModel"
            },
            "nullable": true
          },
          "isFavorite": {
            "type": "boolean"
          },
          "aquiredAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "UserPatchUploadModel": {
        "required": [
          "imageUrl"
        ],
        "type": "object",
        "properties": {
          "userPatchUploadId": {
            "type": "string",
            "format": "uuid"
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "created": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userState": {
            "$ref": "#/components/schemas/UserState"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "username": {
            "type": "string",
            "nullable": true
          },
          "bio": {
            "type": "string",
            "nullable": true
          },
          "profilePictureUrl": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "universityName": {
            "type": "string",
            "nullable": true
          },
          "universityLogoUrl": {
            "type": "string",
            "nullable": true
          },
          "universityProgram": {
            "type": "string",
            "nullable": true
          },
          "created": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "UserRole": {
        "enum": [
          "Unknown",
          "User",
          "PatchMaker",
          "Moderator",
          "Admin"
        ],
        "type": "string"
      },
      "UserState": {
        "enum": [
          "Unknown",
          "Active",
          "Locked",
          "Banned",
          "Deleted"
        ],
        "type": "string"
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Please enter a valid token",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}